Move TomographyResult class to n_qubit_tomography module
Moves TomographyResult definition and plotting helper functions to n_qubit_tomography.py
to avoid circular imports.
