+++ cirq-core/cirq/experiments/qubit_characterizations.py
@@ -17,9 +17,8 @@ from __future__ import annotations
 import dataclasses
 import functools
 import itertools
-import uuid
 from collections.abc import Iterator, Mapping, Sequence
-from typing import Any, cast, Optional, TYPE_CHECKING
+from typing import Any, Optional, TYPE_CHECKING
 
 import attrs
 import numpy as np
@@ -33,10 +32,9 @@ import cirq.vis.histogram as cirq_histogram
 from cirq import circuits, ops, protocols
 from cirq._compat import deprecated
 from cirq.devices import grid_qubit
+from cirq.experiments.n_qubit_tomography import state_tomography, TomographyResult
 
 if TYPE_CHECKING:
-    from mpl_toolkits import mplot3d
-
     import cirq
 
 
@@ -284,80 +282,6 @@ class ParallelRandomizedBenchmarkingResult:
         )
         ax.set_ylabel('Percentile')
         return ax
-
-
-class TomographyResult:
-    """Results from a state tomography experiment."""
-
-    def __init__(self, density_matrix: np.ndarray):
-        """Inits TomographyResult.
-
-        Args:
-            density_matrix: The density matrix obtained from tomography.
-        """
-        self._density_matrix = density_matrix
-
-    @property
-    def data(self) -> np.ndarray:
-        """Returns an n^2 by n^2 complex matrix representing the density
-        matrix of the n-qubit system.
-        """
-        return self._density_matrix
-
-    def plot(
-        self, axes: list[plt.Axes] | None = None, **plot_kwargs: Any
-    ) -> list[plt.Axes]:  # pragma: no cover
-        """Plots the real and imaginary parts of the density matrix as two 3D bar plots.
-
-        Args:
-            axes: A list of 2 `plt.Axes` instances. Note that they must be in
-                3d projections. If not given, a new figure is created with 2
-                axes and the plotted figure is shown.
-            **plot_kwargs: The optional kwargs passed to bar3d.
-
-        Returns:
-            the list of `plt.Axes` being plotted on.
-
-        Raises:
-            ValueError: If axes is a list with length != 2.
-        """
-        show_plot = axes is None
-        if axes is None:
-            fig, axes_v = plt.subplots(1, 2, figsize=(12.0, 5.0), subplot_kw={'projection': '3d'})
-            axes_v = cast(np.ndarray, axes_v)
-            axes = list(axes_v)
-        elif len(axes) != 2:
-            raise ValueError('A TomographyResult needs 2 axes to plot.')
-        mat = self._density_matrix
-        a, _ = mat.shape
-        num_qubits = int(np.log2(a))
-        state_labels = [[0, 1]] * num_qubits
-        kets = []
-        for label in itertools.product(*state_labels):
-            kets.append('|' + str(list(label))[1:-1] + '>')
-        mat_re = np.real(mat)
-        mat_im = np.imag(mat)
-        _matrix_bar_plot(
-            mat_re,
-            r'Real($\rho$)',
-            axes[0],
-            kets,
-            'Density Matrix (Real Part)',
-            ylim=(-1, 1),
-            **plot_kwargs,
-        )
-        _matrix_bar_plot(
-            mat_im,
-            r'Imaginary($\rho$)',
-            axes[1],
-            kets,
-            'Density Matrix (Imaginary Part)',
-            ylim=(-1, 1),
-            **plot_kwargs,
-        )
-        if show_plot:
-            fig.show()
-        return axes
 
 
 @attrs.frozen
@@ -661,13 +585,8 @@ def single_qubit_state_tomography(
     Returns:
         A TomographyResult object that stores and plots the density matrix.
     """
-    from cirq.experiments.n_qubit_tomography import state_tomography
-
     return state_tomography(
-        sampler=sampler,
-        qubits=[qubit],
-        circuit=circuits.Circuit(circuit),
-        repetitions=repetitions,
+        sampler=sampler, qubits=[qubit], circuit=circuits.Circuit(circuit), repetitions=repetitions
     )
 
 
@@ -685,56 +604,6 @@ def two_qubit_state_tomography(
     two qubits before measurements in the z-basis to determine the state
     probabilities $P_{00}, P_{01}, P_{10}.$
 
-    The density matrix rho is decomposed into an operator-sum representation
-    $\sum_{i, j} c_{ij} * \sigma_i \bigotimes \sigma_j$, where $i, j = 0, 1, 2,
-    3$ and $\sigma_0 = I, \sigma_1 = \sigma_x, \sigma_2 = \sigma_y, \sigma_3 =
-    \sigma_z$ are the single-qubit Identity and Pauli matrices.
-
-    Based on the measured probabilities probs and the transformations of the
-    measurement operator by different basis rotations, one can build an
-    overdetermined set of linear equations.
-
-    As an example, if the identity operation (I) is applied to both qubits, the
-    measurement operators are $(I +/- \sigma_z) \bigotimes (I +/- \sigma_z)$.
-    The state probabilities $P_{00}, P_{01}, P_{10}$ thus obtained contribute
-    to the following linear equations (setting $c_{00} = 1$):
-
-    $$
-    \begin{align}
-    c_{03} + c_{30} + c_{33} &= 4*P_{00} - 1 \\
-    -c_{03} + c_{30} - c_{33} &= 4*P_{01} - 1 \\
-    c_{03} - c_{30} - c_{33} &= 4*P_{10} - 1
-    \end{align}
-    $$
-
-    And if a Y/2 rotation is applied to the first qubit and a X/2 rotation
-    is applied to the second qubit before measurement, the measurement
-    operators are $(I -/+ \sigma_x) \bigotimes (I +/- \sigma_y)$. The
-    probabilities obtained instead contribute to the following linear equations:
-
-    $$
-    \begin{align}
-    c_{02} - c_{10} - c_{12} &= 4*P_{00} - 1 \\
-    -c_{02} - c_{10} + c_{12} &= 4*P_{01} - 1 \\
-    c_{02} + c_{10} + c_{12} &= 4*P_{10} - 1
-    \end{align}
-    $$
-
-    Note that this set of equations has the same form as the first set under
-    the transformation $c_{03}$ <-> $c_{02}, c_{30}$ <-> $-c_{10}$ and
-    $c_{33}$ <-> $-c_{12}$.
-
-    Since there are 9 possible combinations of rotations (each producing 3
-    independent probabilities) and a total of 15 unknown coefficients $c_{ij}$,
-    one can cast all the measurement results into a overdetermined set of
-    linear equations numpy.dot(mat, c) = probs. Here c is of length 15 and
-    contains all the $c_{ij}$'s (except $c_{00}$ which is set to 1), and mat
-    is a 27 by 15 matrix having three non-zero elements in each row that are
-    either 1 or -1.
-
-    The least-square solution to the above set of linear equations is then
-    used to construct the density matrix rho.
-
     See Vandersypen and Chuang, Rev. Mod. Phys. 76, 1037 for details and
     Steffen et al, Science 313, 1423 for a related experiment.
 
@@ -748,8 +617,6 @@ def two_qubit_state_tomography(
     Returns:
         A TomographyResult object that stores and plots the density matrix.
     """
-    from cirq.experiments.n_qubit_tomography import state_tomography
-
     return state_tomography(
         sampler=sampler,
         qubits=[first_qubit, second_qubit],
@@ -849,41 +716,6 @@ def _find_inv_matrix(mat: np.ndarray, mat_sequence: np.ndarray) -> int:
     diag_sums = list(np.absolute(np.einsum('...ii->...', mat_prod)))
     idx = diag_sums.index(max(diag_sums))
     return idx
-
-
-def _matrix_bar_plot(
-    mat: np.ndarray,
-    z_label: str,
-    ax: mplot3d.axes3d.Axes3D,
-    kets: Sequence[str] | None = None,
-    title: str | None = None,
-    ylim: tuple[int, int] = (-1, 1),
-    **bar3d_kwargs: Any,
-) -> None:  # pragma: no cover
-    num_rows, num_cols = mat.shape
-    indices = np.meshgrid(range(num_cols), range(num_rows))
-    x_indices = np.array(indices[1]).flatten()
-    y_indices = np.array(indices[0]).flatten()
-    z_indices = np.zeros(mat.size)
-
-    dx = np.ones(mat.size) * 0.3
-    dy = np.ones(mat.size) * 0.3
-    dz = mat.flatten()
-    ax.bar3d(
-        x_indices, y_indices, z_indices, dx, dy, dz, color='#ff0080', alpha=1.0, **bar3d_kwargs
-    )
-
-    ax.set_zlabel(z_label)
-    ax.set_zlim3d(ylim[0], ylim[1])
-
-    if kets is not None:
-        ax.set_xticks(np.arange(num_cols) + 0.15)
-        ax.set_yticks(np.arange(num_rows) + 0.15)
-        ax.set_xticklabels(kets)
-        ax.set_yticklabels(kets)
-
-    if title is not None:
-        ax.set_title(title)
 
 
 def _reduce_gate_seq(
